Uses dpkg-buildpackage for securedrop-app-code
Updates the Ansible-based build logic to use `dpkg-buildpackage`, so
that we can use substvars for Depends, to provide different values for
Trusty/Xenial dependencies.

As such, removes the "sed" implementation to substitute the depends
in-line, which was always a short-term solution anyway.

The fetch logic was updated to find the deb in a different location.

Include `debhelper` in the Dockerfiles used for building. `sudo` also
required, simply because the Ansible logic uses it to run builds,
and it wasn't in the Xenial image by default. Adding to both for
clarity, since it's genuinely a dependency for both.

Installs python-requests for OSSEC URL fetching

The dependency has long been present, but in the Trusty container image
used for building debs, `python-requests` was already installed. That's
not true in the Xenial image, so we'll add it in the build role logic
where it belongs.